# backend/docs/swagger.yaml
basePath: /
definitions:
  kojan-map_shared_models.Report:
    properties:
      deleted:
        type: boolean
      handled:
        type: boolean
      reason:
        type: string
      reportId:
        type: integer
      reportedAt:
        type: string
      reporterGoogleId:
        type: string
      targetPostId:
        type: integer
    type: object
...
  service.PostInfo:
    properties:
      postDate:
        type: string
      postId:
        type: integer
      text:
        type: string
      title:
        type: string
      userId:
        type: string
    type: object
  service.ReportDetailResponse:
    properties:
      deleted:
        type: boolean
      handled:
        type: boolean
      post:
        allOf:
        - $ref: '#/definitions/service.PostInfo'
        description: Target post details
      reason:
        type: string
      reportId:
        type: integer
      reportedAt:
        type: string
      reporterGoogleId:
        type: string
      targetPostId:
        type: integer
    type: object
  service.ReportListResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      reports:
        items:
          $ref: '#/definitions/kojan-map_shared_models.Report'
        type: array
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@kojan-map.example.com
    name: API サポート
  description: |-
    こじゃんとやまっぷのバックエンドAPIドキュメント

    このAPIは一般ユーザー、ビジネスユーザー、管理者向けの機能を提供します。
    認証が必要なエンドポイントではBearer JWTトークンを使用します。
  title: こじゃんとやまっぷ API
  version: "1.0"
paths:
...
  /api/admin/reports:
    get:
      consumes:
      - application/json
      description: ページネーション付きで通報一覧を取得します
      parameters:
      - default: 1
        description: ページ番号
        in: query
        name: page
        type: integer
      - default: 20
        description: 1ページあたりの件数
        in: query
        name: pageSize
        type: integer
      - description: 処理済みフィルター
        in: query
        name: handled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 通報一覧
          schema:
            $ref: '#/definitions/service.ReportListResponse'
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通報一覧を取得
      tags:
      - Admin Reports
  /api/admin/reports/{id}:
    get:
      consumes:
      - application/json
      description: 指定したIDの通報詳細と対象投稿情報を取得します
      parameters:
      - description: 通報ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 通報詳細
          schema:
            $ref: '#/definitions/service.ReportDetailResponse'
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 通報が見つからない
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通報詳細を取得
      tags:
      - Admin Reports
  /api/admin/reports/{id}/handle:
    put:
      consumes:
      - application/json
      description: 指定したIDの通報を処理済み状態に更新します
      parameters:
      - description: 通報ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 処理成功
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通報を処理済みにする
      tags:
      - Admin Reports
...
  /api/admin/users/{userId}:
    delete:
      consumes:
      - application/json
      description: 指定したユーザーIDのユーザーを削除します
      parameters:
      - description: ユーザーID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 削除成功
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: 不正なリクエスト
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ユーザーを削除
      tags:
      - Admin Users
...
  /api/users:
    get:
      consumes:
      - application/json
      description: ページネーション付きで登録ユーザーの一覧を取得します
      parameters:
      - default: 1
        description: ページ番号
        in: query
        name: page
        type: integer
      - default: 20
        description: 1ページあたりの件数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ユーザー一覧とページネーション情報
          schema:
            additionalProperties: true
            type: object
        "500":
          description: サーバーエラー
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ユーザー一覧を取得
      tags:
      - Admin Users
...
securityDefinitions:
  BearerAuth:
    description: JWT認証トークン。ヘッダーに "Bearer {token}" の形式で指定してください。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"